cvs studio/speech_text_image_converter/requirements.txt
SpeechRecognition==3.10.0
pydub==0.25.1
pytesseract==0.3.10
Pillow==10.0.0
fpdf==1.7.2
numpy==1.24.3
transformers==4.30.2
torch==2.0.1
torchaudio==2.0.2
opencv-python==4.8.0.74
pdf2image==1.16.3
python-docx==0.8.11
librosa==0.10.0 